refactor: LatestAuctionResults section
